Added documentation for the Meal, MealPlanActivity, and MealPlanFragment classes. Removed MealTestFragment and transferred its functionality to MealPlanFragment
