+++ app/src/main/java/com/example/foodverse/MealPlanFragment.java
@@ -9,6 +9,7 @@ import android.os.Bundle;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
+import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.Button;
 import android.widget.DatePicker;
@@ -40,39 +41,43 @@ import java.util.HashSet;
 import java.util.Set;
 
 /**
- * MealPlanFragment
- * Currently not a functional class.
- * This
+ * MealTestFragment
+ * Used to create a fragment for adding, editing or deleting meals.
+ * Originally this was created as a test, but it has since supplanted
+ * the original MealPlanFragment class.
  */
 
-public class MealPlanFragment extends DialogFragment  {
+public class MealTestFragment extends DialogFragment implements AdapterView.OnItemSelectedListener {
     private Meal meal;
     private EditText date;
     // Recipe-related elements still need to be implemented
     private ListView ingredientList;
     private EditText servings;
 
-    private Button mealDate;
-    private Button addButton;
+    private Button mealDate; // Date selector
+    private Button addButton; // Plus button to add ingredient to meal
     private Date date2;
-    private Spinner recipeSpinner;
-    private Spinner ingredientSpinner;
+    private Spinner recipeSpinner; // Spinner for recipes
+    private Spinner ingredientSpinner; // Spinner for ingredients
     private ArrayList<Ingredient> mealIngredients = new ArrayList<>();
-    private ArrayList<Ingredient> addedIngredients = new ArrayList<>();
-    private ArrayList<String> listedIngredients = new ArrayList<>();
+    private ArrayList<String> ingredientStringList = new ArrayList<>();
+    private ArrayList<Ingredient> arraySpinner = new ArrayList<>();
     private ArrayAdapter<String> ingAdapter;
     private FirebaseFirestore db;
     private CollectionReference ingRef, storedRef;
     private HashSet<Ingredient> set = new HashSet<>();
+    private ArrayAdapter<Ingredient> listViewAdapter;
+    private boolean ingFinished = false, storedFinished = false;
+    private MealPlanActivity act;
 
-    private MealPlanFragment.OnFragmentInteractionListener listener;
+    private MealTestFragment.OnFragmentInteractionListener listener;
 
-    public MealPlanFragment() {
+    public MealTestFragment() {
         super();
         this.meal = null;
     }
 
-    public MealPlanFragment(Meal meal) {
+    public MealTestFragment(Meal meal) {
         super();
         this.meal = meal;
     }
@@ -86,8 +91,8 @@ public class MealPlanFragment extends DialogFragment  {
     @Override
     public void onAttach(Context context) {
         super.onAttach(context);
-        if (context instanceof MealPlanFragment.OnFragmentInteractionListener) {
-            listener = (MealPlanFragment.OnFragmentInteractionListener) context;
+        if (context instanceof MealTestFragment.OnFragmentInteractionListener) {
+            listener = (MealTestFragment.OnFragmentInteractionListener) context;
         } else {
             throw new RuntimeException(context.toString()
                     + "must implement OnFragmentInteractionListener");
@@ -104,9 +109,14 @@ public class MealPlanFragment extends DialogFragment  {
         recipeSpinner = view.findViewById(R.id.recipe_spinner);
         ingredientSpinner = view.findViewById(R.id.meal_ingredient_spinner);
         addButton = view.findViewById(R.id.add_meal_ingredient_button);
-        ingredientList = view.findViewById(R.id.meal_list);
+        ingredientList = view.findViewById(R.id.meal_fragment_list);
+        listViewAdapter = new IngredientAdapter(getActivity(), mealIngredients);
+        ingredientList.setAdapter(listViewAdapter);
 
-        ArrayAdapter<String> listedAdapter = new ArrayAdapter<String>(getActivity(), R.layout.content_stored_ingredient, listedIngredients);
+        ingredientSpinner.setOnItemSelectedListener(this);
+
+
+        //ArrayAdapter<String> listedAdapter = new ArrayAdapter<String>(getActivity(), R.layout.content_stored_ingredient, listedIngredients);
         //ingredientList.setAdapter(listedAdapter);
 
         DatePickerDialog.OnDateSetListener dateSetListener;
@@ -120,56 +130,67 @@ public class MealPlanFragment extends DialogFragment  {
             }
         };
 
+        ingAdapter = new ArrayAdapter<String>(getActivity(), R.layout.ingredient_spinner, ingredientStringList);
         db = FirebaseFirestore.getInstance();
         db.enableNetwork();
-        ingRef = db.collection("Ingredients");
-        storedRef = db.collection("StoredIngredients");
+//        ingRef = db.collection("Ingredients");
+//        storedRef = db.collection("StoredIngredients");
+//
+//        ingRef.addSnapshotListener(new EventListener<QuerySnapshot>() {
+//            @Override
+//            public void onEvent(@Nullable QuerySnapshot queryDocumentSnapshots, @Nullable
+//                    FirebaseFirestoreException error) {
+//                // Add ingredients from the cloud
+//                for(QueryDocumentSnapshot doc: queryDocumentSnapshots) {
+//                    String hashCode = doc.getId();
+//                    String description = (String) doc.getData().get("Description");
+//                    Log.d("MEALFRAG", description);
+//                    Long count = (Long) doc.getData().get("Count");
+//                    Ingredient ing = new Ingredient(description, count.intValue());
+//                    if (!set.contains(ing)) {
+//                        arraySpinner.add(ing);
+//                        set.add(ing);
+//                        Log.d("MEALFRAG", "Added ing");
+//                    }
+//                }
+//                ingFinished = true;
+//                if (ingFinished && storedFinished) {
+//                    populateSpinner();
+//                }
+//            }
+//        });
+//
+//        storedRef.addSnapshotListener(new EventListener<QuerySnapshot>() {
+//            @Override
+//            public void onEvent(@Nullable QuerySnapshot queryDocumentSnapshots, @Nullable
+//                    FirebaseFirestoreException error) {
+//                // Add ingredients from the cloud
+//                for(QueryDocumentSnapshot doc: queryDocumentSnapshots) {
+//                    String hashCode = doc.getId();
+//                    String description = (String) doc.getData().get("Description");
+//                    Log.d("MEALFRAG", description);
+//                    Long count = (Long) doc.getData().get("Count");
+//                    Ingredient ing = new Ingredient(description, count.intValue());
+//                    if (!set.contains(ing)) {
+//                        arraySpinner.add(ing);
+//                        set.add(ing);
+//                        Log.d("MEALFRAG", "Added ing");
+//                    }
+//                }
+//                storedFinished = true;
+//                if (ingFinished && storedFinished) {
+//                    populateSpinner();
+//                }
+//            }
+//        });
+        act = (MealPlanActivity) getActivity();
+        for (int i = 0; i < act.getDatabaseIngredients().size(); i++) {
+            ingredientStringList.add(
+                    act.getDatabaseIngredients().get(i).getDescription());
+        }
 
-        ArrayList<String> arraySpinner = new ArrayList<>();
-        arraySpinner.add("Add New Ingredient");
-
-        ingRef.addSnapshotListener(new EventListener<QuerySnapshot>() {
-            @Override
-            public void onEvent(@Nullable QuerySnapshot queryDocumentSnapshots, @Nullable
-                    FirebaseFirestoreException error) {
-                // Add ingredients from the cloud
-                for(QueryDocumentSnapshot doc: queryDocumentSnapshots) {
-                    String hashCode = doc.getId();
-                    String description = (String) doc.getData().get("Description");
-                    Log.d("MEALFRAG", description);
-                    Long count = (Long) doc.getData().get("Count");
-                    Ingredient ing = new Ingredient(description, count.intValue());
-                    if (!set.contains(ing)) {
-                        arraySpinner.add(description);
-                        set.add(ing);
-                    }
-                }
-            }
-        });
-
-        storedRef.addSnapshotListener(new EventListener<QuerySnapshot>() {
-            @Override
-            public void onEvent(@Nullable QuerySnapshot queryDocumentSnapshots, @Nullable
-                    FirebaseFirestoreException error) {
-                // Clear the old list
-                mealIngredients.clear();
-                // Add ingredients from the cloud
-                for(QueryDocumentSnapshot doc: queryDocumentSnapshots) {
-                    String hashCode = doc.getId();
-                    String description = (String) doc.getData().get("Description");
-                    Log.d("MEALFRAG", description);
-                    Long count = (Long) doc.getData().get("Count");
-                    Ingredient ing = new Ingredient(description, count.intValue());
-                    if (!set.contains(ing)) {
-                        arraySpinner.add(description);
-                        set.add(ing);
-                    }
-                }
-            }
-        });
-
-        ingAdapter = new IngredientSpinnerAdapter(getActivity(), arraySpinner);
         ingAdapter.setDropDownViewResource(R.layout.ingredient_spinner);
+
         ingredientSpinner.setAdapter(ingAdapter);
 
 
@@ -178,34 +199,29 @@ public class MealPlanFragment extends DialogFragment  {
         int year = calendar.get(Calendar.YEAR);
         int month = calendar.get(Calendar.MONTH);
         int day = calendar.get(Calendar.DAY_OF_MONTH);
-        //setNewExpiryDate(calendar);
+        setNewExpiryDate(calendar);
 
         if (meal != null) {
             // Set the values for the various fields
-            //ingredientDescription.setText(ingredient.getDescription());
-            //ingredientCount.setText(Integer.toString(ingredient.getCount()));
-            //ingredientCost.setText(Integer.toString(ingredient.getUnitCost()));
 
             // Load the time information
-            //calendar.setTime(ingredient.getBestBefore());
+            calendar.setTime(meal.getDate());
             year = calendar.get(Calendar.YEAR);
             month = calendar.get(Calendar.MONTH);
             day = calendar.get(Calendar.DAY_OF_MONTH);
             setNewExpiryDate(calendar);
 
+            //mealIngredients = meal.getIngredients();
+
+            for (int i = 0; i < meal.getIngredients().size(); i++) {
+                mealIngredients.add(meal.getIngredients().get(i));
+            }
+            listViewAdapter.notifyDataSetChanged();
         }
 
         DatePickerDialog datePickerDialog = new DatePickerDialog(
                 this.getActivity(), AlertDialog.THEME_HOLO_LIGHT,
                 dateSetListener, year, month, day);
-
-
-        //ingredientExpiry.setOnClickListener(new View.OnClickListener() {
-        //    @Override
-        //    public void onClick(View view) {
-        //        datePickerDialog.show();
-        //    }
-        //});
 
         mealDate.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -214,8 +230,6 @@ public class MealPlanFragment extends DialogFragment  {
             }
         });
 
-        //ArrayList<String> listedIngredients = new ArrayList<>();
-
 
 
         // When the user clicks on the plus button to add an ingredient
@@ -223,35 +237,43 @@ public class MealPlanFragment extends DialogFragment  {
             @Override
             public void onClick(View view) {
                 //addedIngredients.add((Ingredient) ingredientSpinner.getSelectedItem());
-                listedIngredients.add((String) ingredientSpinner.getSelectedItem());
-                listedAdapter.notifyDataSetChanged();
+                //listedIngredients.add((String) ingredientSpinner.getSelectedItem());
+                //listedAdapter.notifyDataSetChanged();
+                int ingIndex;
+                Log.d("MealFrag", "Adding ingredient");
+                ingIndex = ingredientSpinner.getSelectedItemPosition();
+                mealIngredients.add(act.getDatabaseIngredients().get(ingIndex));
+                //ingAdapter.notifyDataSetChanged();
+                listViewAdapter.notifyDataSetChanged();
             }
         });
-
-        //IngredientAdapter addedAdapter = new IngredientAdapter(getActivity(), addedIngredients);
-        //ingredientList.setAdapter(listedAdapter);
-
-        //ArrayList<String> listedIngredients = new ArrayList<String>();
 
         AlertDialog.Builder builder = new AlertDialog.Builder(getContext());
         return builder
                 .setView(view)
                 .setTitle("Add/Edit Meal")
                 .setNeutralButton("Cancel", null)
-                .setNegativeButton("Delete", null)
+                .setNegativeButton("Delete", (dialog, which) -> {
+                    listener.mealDeleted();
+                })
                 .setPositiveButton("Confirm", new DialogInterface.OnClickListener() {
                     @RequiresApi(api = Build.VERSION_CODES.O)
                     @Override
                     public void onClick(DialogInterface dialog, int which) {
-                        //String city = cityName.getText().toString();
-                        //String province = provinceName.getText().toString();
                         //listener.onOkPressed(new City(city, province));
-                        Meal meal = new Meal();
-                        LocalDate date = date2.toInstant().atZone(ZoneId.systemDefault()).toLocalDate();
-                        meal.setDate(date2);
+                        Meal newMeal = new Meal();
+                        //LocalDate date = date2.toInstant().atZone(ZoneId.systemDefault()).toLocalDate();
+                        newMeal.setDate(date2);
                         // Add Recipe Stuff
                         // Add array of ingredients to meal
-                        meal.setIngredients(mealIngredients);
+                        newMeal.setIngredients(mealIngredients);
+                        if (meal == null) {
+                            listener.mealAdded(newMeal);
+                        } else {
+                            listener.mealEdited(newMeal);
+                        }
+                        //listener.mealAdded(newMeal);
+
                     }
                 }).create();
     }
@@ -278,6 +300,16 @@ public class MealPlanFragment extends DialogFragment  {
         String date = Integer.toString(year) + "-" + monthStr + "-" + dayStr;
         mealDate.setText(date);
         date2 = calendar.getTime();
+    }
+
+    @Override
+    public void onItemSelected(AdapterView<?> parent, View view, int pos, long id) {
+        //this.locationInt = pos;
+        //ingredientSpinner.setSelection(pos);
+    }
+
+    @Override
+    public void onNothingSelected(AdapterView<?> parent) {
     }
 
 }
